[Refactor] 미션 전 프로젝트 구조 변경진행
`+` Domain 별 패키지 구성
