+++ src/main/java/subway/station/dto/request/StationRequest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station.dto.request;
 
 public class StationRequest {
     private String name;
